generator updated

+++ src/benchmark.py
@@ -14,7 +14,7 @@ def benchmark():
     GENERATE_LEN = 50
 
     device = torch.device("cuda")
-    model = TinyLlama(VOCAB_SIZE, HIDDEN_DIM, NUM_HEADS, NUM_LAYERS, MAX_SEQ_LEN).to(device)
+    model = TinyLlama(VOCAB_SIZE, HIDDEN_DIM, NUM_HEADS, NUM_LAYERS, MAX_SEQ_LEN).to(device).half()
     model.eval()
 
     prompt = torch.randint(0, VOCAB_SIZE, (BATCH_SIZE, PROMPT_LEN), device=device)
